test: Fixture & Test Method 분리.
