+++ src/test/java/nextstep/subway/auth/acceptance/AuthAcceptanceTest.java
@@ -1,36 +1,16 @@
 package nextstep.subway.auth.acceptance;
 
-import static nextstep.subway.member.MemberAcceptanceTest.회원_생성을_요청;
+import static nextstep.subway.auth.acceptance.AuthFixture.로그인_요청;
+import static nextstep.subway.member.MemberFixture.회원_생성을_요청;
 
-import io.restassured.RestAssured;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
-import nextstep.subway.auth.dto.TokenRequest;
 import nextstep.subway.auth.dto.TokenResponse;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.http.MediaType;
 
 public class AuthAcceptanceTest extends AcceptanceTest {
 
-    public static ExtractableResponse<Response> 로그인_요청(String email, String password) {
-        TokenRequest tokenRequest = new TokenRequest(email, password);
-
-        return RestAssured
-            .given().log().all()
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body(tokenRequest)
-            .when().post("/login/token")
-            .then().log().all()
-            .extract();
-    }
-
-    public static String 로그인_요청_후_토큰_가져오기(String email, String password) {
-        TokenResponse tokenResponse = 로그인_요청(email, password).as(TokenResponse.class);
-        return tokenResponse.getAccessToken();
-    }
 
     /*
     Feature: 로그인 기능
