+++ src/test/java/nextstep/subway/auth/acceptance/AuthFixture.java
@@ -1,0 +1,28 @@
+package nextstep.subway.auth.acceptance;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.subway.auth.dto.TokenRequest;
+import nextstep.subway.auth.dto.TokenResponse;
+import org.springframework.http.MediaType;
+
+public class AuthFixture {
+
+    public static ExtractableResponse<Response> 로그인_요청(String email, String password) {
+        TokenRequest tokenRequest = new TokenRequest(email, password);
+
+        return RestAssured
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(tokenRequest)
+            .when().post("/login/token")
+            .then().log().all()
+            .extract();
+    }
+
+    public static String 로그인_요청_후_토큰_가져오기(String email, String password) {
+        TokenResponse tokenResponse = 로그인_요청(email, password).as(TokenResponse.class);
+        return tokenResponse.getAccessToken();
+    }
+}
