+++ src/test/java/nextstep/subway/path/PathAcceptanceTest.java
@@ -1,8 +1,8 @@
 package nextstep.subway.path;
 
-import static nextstep.subway.line.acceptance.LineAcceptanceTest.지하철_노선_등록되어_있음;
-import static nextstep.subway.line.acceptance.LineSectionAcceptanceTest.지하철_노선에_지하철역_등록_요청;
-import static nextstep.subway.station.StationAcceptanceTest.지하철역_등록되어_있음;
+import static nextstep.subway.line.acceptance.LineFixture.지하철_노선_등록되어_있음;
+import static nextstep.subway.line.acceptance.LineSectionFixture.지하철_노선에_지하철역_등록_요청;
+import static nextstep.subway.station.StationFixture.지하철역_등록되어_있음;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.restassured.RestAssured;
