+++ src/test/java/nextstep/subway/path/PathFixture.java
@@ -1,0 +1,5 @@
+package nextstep.subway.path;
+
+public class PathFixture {
+
+}
